backend/requirements.txt
aiosql==3.3.1
alembic==1.7.6
anyio==3.5.0
asgiref==3.5.0
asyncpg==0.25.0
bcrypt==3.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.11
click==8.0.3
colorama==0.4.4; platform_system == "Windows"
colorama==0.4.4; sys_platform == "win32"
contextlib2==21.6.0
databases==0.5.5
dnspython==2.2.0
email-validator==1.1.3
fastapi==0.73.0
greenlet==1.1.2; python_version >= "3" and (platform_machine == "aarch64" or platform_machine == "ppc64le" or platform_machine == "x86_64" or platform_machine == "amd64" or platform_machine == "AMD64" or platform_machine == "win32" or platform_machine == "WIN32")
gunicorn==20.1.0
h11==0.12.0
idna==3.3
loguru==0.6.0
mako==1.1.6
markupsafe==2.0.1
passlib==1.7.4
psycopg2-binary==2.9.3
pycparser==2.21
pydantic==1.9.0
pyjwt==2.3.0
pypika==0.48.8
python-dotenv==0.19.2
python-slugify==5.0.2
requests==2.28.0
six==1.16.0
sniffio==1.2.0
sqlalchemy==1.4.31
starlette==0.17.1
text-unidecode==1.3
typing-extensions==3.10.0.2
unidecode==1.3.2
urllib3==1.26.9
uvicorn==0.17.4
win32-setctime==1.1.0; sys_platform == "win32"
openai==0.19.0